requirements.txt
streamlit
streamlit-pdf-viewer
python-dotenv
langchain-google-genai
langchain_community 
beautifulsoup4
pypdf
fpdf
# chromadb==1.0.13